+++ src/main/java/br/edu/ifsp/spo/bike_integration/repository/EmailRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.edu.ifsp.spo.bike_integration.model.Email;
 
-public interface EmailRepository extends JpaRepository<Email, Integer>{
+public interface EmailRepository extends JpaRepository<Email, Long>{
 	Optional<Email> findByValor(String email);
 }
