+++ src/main/java/br/edu/ifsp/spo/bike_integration/repository/PessoaFisicaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.edu.ifsp.spo.bike_integration.model.PessoaFisica;
 
-public interface PessoaFisicaRepository extends JpaRepository<PessoaFisica, Integer>{
+public interface PessoaFisicaRepository extends JpaRepository<PessoaFisica, Long>{
 	Optional<PessoaFisica> findByCpf(String cpf);
 }
